requirements.txt
asgiref==3.7.2
Babel==2.12.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-import-export==3.2.0
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
phonenumbers==8.13.15
Pillow==9.5.0
pycop==0.0.11
python-dotenv==1.0.0
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.6.3
xlrd==2.0.1
xlwt==1.3.0
